add user data in document preview dialog to handle db provider embedd

+++ apps/web/src/components/document-preview-dialog.tsx
@@ -15,8 +15,14 @@ import { Button, buttonVariants } from './ui/button'
 import { Input } from './ui/input'
 import { Textarea } from './ui/textarea'
 import { Label } from './ui/label'
+import { Badge } from './ui/badge'
 
 type Props = {
+  user: {
+    id: string
+    dbStatus: boolean
+    vectorProvider: 'pinecone' | 'supabase' | null
+  }
   document: DocumentProps
   children: React.ReactNode
 }
@@ -27,49 +33,64 @@ const parseContent = (content: DocumentProps['content']) => {
     return chunk.pageContent ? (acc + chunk.pageContent) : acc
   }, '')
     .trim()
-    .replace(/\n/g, ' ')
 }
 
-function DocumentPreviewDialog ({ document, children }: Props) {
+function DocumentPreviewDialog ({ user, document, children }: Props) {
+  const [isOpen, setIsOpen] = useState(false)
   const [name, setName] = useState(document.name)
 
   const handleEmbedding = async () => {
-    await fetch('/api/embeddings/supabase', { // use vector DB selected for the user
+    if (!user.dbStatus || !user.vectorProvider) return alert('You must connect to a database first')
+    await fetch(`/api/embeddings/${user.vectorProvider}`, {
       method: 'POST',
       body: JSON.stringify({
+        name: name || document.name,
+        userId: document.user.id,
+        docId: document.id,
         content: document.content
       })
     })
+
+    setIsOpen(false)
   }
 
   const handleDeleteEmbedding = async () => {
-    await fetch('/api/embeddings/supabase', { // use vector DB selected for the user
+    if (!user.dbStatus || !user.vectorProvider) return alert('You must connect to a database first')
+    await fetch(`/api/embeddings/${user.vectorProvider}`, {
       method: 'DELETE',
       body: JSON.stringify({
+        userId: document.user.id,
         docId: document.id,
-        ids: document.supaEmbeddedIds // only for supabase vector DB
+        ids: document.embeddedIds
       })
     })
+
+    setIsOpen(false)
   }
 
   return (
-    <Dialog onOpenChange={(open) => {
-      if (!open) {
-        setName(document.name)
-      }
-    }}>
+    <Dialog
+      open={isOpen}
+      onOpenChange={(open) => {
+        setIsOpen(open)
+        if (!open) {
+          setName(document.name)
+        }
+      }}
+    >
       <DialogTrigger asChild>
         {children}
       </DialogTrigger>
       <DialogContent className="sm:max-w-4xl">
         <DialogHeader>
-          <DialogTitle>
+          <DialogTitle className='flex items-center'>
             {document.name}
+            <Badge className='ml-2' variant='secondary'>{document.type}</Badge>
           </DialogTitle>
           <DialogDescription>
             {document.isTrained
               ? (
-              <>Your document is ready to be used by an agent.</>
+                <>Your document is ready to be used by an agent.</>
                 )
               : <>You must train this file before you can use it in an agent.</>}
           </DialogDescription>
@@ -99,8 +120,8 @@ function DocumentPreviewDialog ({ document, children }: Props) {
         <DialogFooter className="!justify-between">
           {document.isTrained
             ? <Button onClick={handleDeleteEmbedding} variant='destructive'>
-            Delete
-          </Button>
+              Delete
+            </Button>
             : <div />}
           <div className="sm:space-x-2">
             <DialogClose className={cn(buttonVariants({ variant: 'secondary' }))}>
@@ -108,14 +129,14 @@ function DocumentPreviewDialog ({ document, children }: Props) {
             </DialogClose>
             {document.isTrained
               ? (
-              <Button className="bg-green-700 text-white hover:bg-green-700/90">
-                Save Changes
-              </Button>
+                <Button className="bg-green-700 text-white hover:bg-green-700/90">
+                  Save Changes
+                </Button>
                 )
               : (
-              <Button className="bg-green-700 text-white hover:bg-green-700/90" onClick={handleEmbedding}>
-                Generate Embeddings
-              </Button>
+                <Button className="bg-green-700 text-white hover:bg-green-700/90" onClick={handleEmbedding}>
+                  Generate Embeddings
+                </Button>
                 )}
           </div>
         </DialogFooter>
